+++ src/NovelSpider.Entity/Properties/AssemblyInfo.cs
@@ -16,7 +16,7 @@ using System.Runtime.Versioning;
 [assembly: AssemblyTrademark("")]
 [assembly: ComVisible(false)]
 [assembly: Guid("9947a03f-65fb-461a-b305-b0ce4a61f0e4")]
-[assembly: AssemblyFileVersion("10.1.0.0")]
-[assembly: AssemblyVersion("10.1.0.0")]
+[assembly: AssemblyFileVersion("10.1.1.0")]
+[assembly: AssemblyVersion("10.1.1.0")]